--- Live_page/src/live-drawing/chat-components/index.css
@@ -1,47 +1,0 @@
-.chatScreen{
-    position: relative;
-    border: gray 1px solid;
-    overflow-y: auto;
-    height: 90%;
-    padding-top: 1rem;
-}
-
-.message{
-    text-align: left;
-    margin-left: 5px;
-    margin-bottom: 10px;
-}
-
-.message>p{
-    margin: 2px;
-}
-
-.admin .chat-text {
-    border: gray 1px solid !important;
-}
-
-.chat-text {
-    border:lightskyblue 1px solid;
-    border-radius: 20px;
-    display: inline-block;
-    padding: 5px 10px 5px 10px;
-}
-
-.message>.userName{
-    font-size: 12px;
-}
-
-.message .admin {
-    text-align: center;
-}
-
-
-.message.my{
-    text-align: right;
-    margin-right: 5px;
-}
-
-.input-container {
-    /* border: black 1px solid; */
-    overflow: auto;
-}
