+++ Live_page/src/live-drawing/chat-components/interfaces/chat-component-props-interface.ts
@@ -1,4 +1,6 @@
+import { RoomInfo } from '../../interfaces/socket-interfaces';
+
 export interface ChatComponentProps {
-  userName: string | null;
+  roomInfo: RoomInfo;
   socket: SocketIOClient.Socket | null;
 }
