+++ Live_page/src/live-drawing/index.tsx
@@ -3,7 +3,6 @@ import { useParams } from 'react-router-dom';
 
 import SidebarComponent from './sidebar-components';
 import DrawComponent from './draw-components';
-import ChatComponent from './chat-components';
 
 import Swal from 'sweetalert2';
 import withReactContent from 'sweetalert2-react-content';
@@ -48,25 +47,25 @@ function LiveDrawing() {
       userId: '47a064dd-ab37-4990-aef8-cca398b24b2b',
       userImage:
         'https://lh3.googleusercontent.com/a-/AOh14Ggzk1sZlAeI4hnd0bZyYd7yS1Nqq04glqSQlywLpg=s96-c',
-      userName: '김용욱',
+      username: '김용욱',
     },
     {
       userId: '73dd814c-8026-459b-a2ed-4863ddb79750',
       userImage:
         'https://lh6.googleusercontent.com/-h-I_zB0DmFk/AAAAAAAAAAI/AAAAAAAAAAA/AMZuucnE3WNy2c7s-MYKHpFFSQgdcAifNg/s96-c/photo.jpg',
-      userName: '김유석',
+      username: '김유석',
     },
     {
       userId: 'b02fcfc9-c8d2-4900-8b8b-a7e2b3fa342c',
       userImage:
         'https://lh5.googleusercontent.com/-0XZwgDcb5yU/AAAAAAAAAAI/AAAAAAAAAAA/AMZuucnyQd6S3wBQrFAPoAVUkdUO3q6nMA/s96-c/photo.jpg',
-      userName: 'M K',
+      username: 'M K',
     },
     {
       userId: '776a10b4-03e7-455c-88d4-f9f908e9b846',
       userImage:
         'https://lh5.googleusercontent.com/-UD1QQESYljk/AAAAAAAAAAI/AAAAAAAAAAA/AMZuucntz6Rz06XZSwFKdXXnwkw2u24Ahw/s96-c/photo.jpg',
-      userName: '김형우',
+      username: '김형우',
     },
   ];
 
@@ -84,25 +83,25 @@ function LiveDrawing() {
       {
         roomId: 'feda6c99-f05a-4fdd-88f6-fa13de0a9e12',
         socketId: null,
-        userName: '김용욱',
+        username: '김용욱',
         userId: '47a064dd-ab37-4990-aef8-cca398b24b2b',
       },
       {
         roomId: 'feda6c99-f05a-4fdd-88f6-fa13de0a9e12',
         socketId: null,
-        userName: '김유석',
+        username: '김유석',
         userId: '73dd814c-8026-459b-a2ed-4863ddb79750',
       },
       {
         roomId: 'feda6c99-f05a-4fdd-88f6-fa13de0a9e12',
         socketId: null,
-        userName: 'M K',
+        username: 'M K',
         userId: 'b02fcfc9-c8d2-4900-8b8b-a7e2b3fa342c',
       },
       {
         roomId: 'feda6c99-f05a-4fdd-88f6-fa13de0a9e12',
         socketId: null,
-        userName: '김형우',
+        username: '김형우',
         userId: '776a10b4-03e7-455c-88d4-f9f908e9b846',
       },
     ],
@@ -217,6 +216,7 @@ function LiveDrawing() {
         layers={layers}
         roomInfo={roomInfo}
         topLayer={topLayer}
+        socket={socket}
         userProfileInfos={userProfileInfos}
         setTopLayer={setTopLayer}
         setIsLiveClosed={setIsLiveClosed}
@@ -232,10 +232,6 @@ function LiveDrawing() {
         setIsLiveClosed={setIsLiveClosed}
         setLayers={setLayers}
       />
-      {/* <ChatComponent
-        userName={roomInfo.userName}
-        socket={socket}
-      ></ChatComponent> */}
     </>
   );
 }
