+++ Live_page/src/live-drawing/interfaces/sidebar-components-interfaces.ts
@@ -6,6 +6,7 @@ export interface SidebarComponentProps {
   layers: Layer[];
   roomInfo: RoomInfo;
   topLayer: Layer | null;
+  socket: SocketIOClient.Socket | null;
   userProfileInfos: UserProfileInfo[];
   setTopLayer: React.Dispatch<React.SetStateAction<Layer | null>>;
   setIsLiveClosed: React.Dispatch<React.SetStateAction<boolean>>;
