+++ Live_page/src/live-drawing/interfaces/socket-interfaces.ts
@@ -7,7 +7,7 @@ export interface RoomInfo {
 }
 
 export interface UserProfileInfo {
-  userName: string | null;
+  username: string | null;
   userId: string;
   userImage: string;
 }
@@ -17,7 +17,7 @@ export interface ResponseData {
 }
 
 export interface UserInfo {
-  userName: string;
+  username: string;
   socketId: SocketIOClient.Socket | null;
   userId: string;
   roomId: string;
