+++ Live_page/src/live-drawing/sidebar-components/components/UserSelectButtonComponent.tsx
@@ -3,8 +3,6 @@ import { Layer } from '../../interfaces/draw-components-interfaces';
 import { UserSelectButtonComponentProps } from '../interfaces/user-select-button-interfaces';
 import { UserProfileInfo } from '../../interfaces/socket-interfaces';
 
-import '../index.css';
-
 function UserSelectButtonComponent(props: UserSelectButtonComponentProps) {
   return (
     <>
@@ -27,8 +25,8 @@ function UserSelectButtonComponent(props: UserSelectButtonComponentProps) {
                   }
                   ${
                     layer.canvasId === props.topLayer?.canvasId
-                      ? ' bg-blue-500 text-white '
-                      : ''
+                      ? ' bg-blue-500 text-white shadow-inner '
+                      : ' hover:bg-gray-200'
                   }
                 `}
               >
@@ -44,13 +42,13 @@ function UserSelectButtonComponent(props: UserSelectButtonComponentProps) {
                 <img
                   className={`w-12 h-12 rounded-full my-2 ${
                     layer.canvasId === props.topLayer?.canvasId
-                      ? 'ring-2 ring-white'
+                      ? 'ring-2 ring-white shadow-lg'
                       : ''
                   }`}
                   src={`${user?.userImage}`}
-                  alt={`${user?.userName}`}
+                  alt={`${user?.username}`}
                 />
-                <p>{`${user?.userName}`}</p>
+                <p>{`${user?.username}`}</p>
               </div>
             );
         })}
