+++ Live_page/src/live-drawing/sidebar-components/index.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import UserSelectButtonComponent from './components/UserSelectButtonComponent';
 import CloseButtonComponent from './components/CloseButtonComponent';
 import VoiceChatComponent from '../voice-components/';
+import ChatComponent from '../chat-components';
 
 import { SidebarComponentProps } from '../interfaces/sidebar-components-interfaces';
 
@@ -9,9 +10,7 @@ function SidebarComponent(props: SidebarComponentProps) {
   return (
     <div className='flex-initial w-20 flex flex-col border-r divide-y'>
       <div className='h-20 flex flex-wrap justify-center content-center'>
-        <div className=''>
-          <i className='ri-2x ri-leaf-fill text-blue-500'></i>
-        </div>
+        <i className='ri-2x ri-leaf-fill text-blue-500'></i>
       </div>
       <UserSelectButtonComponent
         topLayer={props.topLayer}
@@ -22,7 +21,7 @@ function SidebarComponent(props: SidebarComponentProps) {
       />
       <div className='flex-grow'></div>
       <VoiceChatComponent />
-      <p>유저, 채팅 버튼</p>
+      <ChatComponent roomInfo={props.roomInfo} socket={props.socket} />
       <div className='flex-grow'></div>
       <CloseButtonComponent
         roomInfo={props.roomInfo}
