--- Live_page/src/live-drawing/voice-components/interfaces/voice-component-props-interface.ts
@@ -1,8 +1,0 @@
-import { IAgoraRTCClient, ILocalAudioTrack } from 'agora-rtc-sdk-ng';
-
-export interface VoiceToolComponentProps {
-  client: IAgoraRTCClient | null;
-  localAudioTrack: ILocalAudioTrack | null;
-  speaker: boolean | null;
-  mic: boolean | null;
-}
